[Mentions] Minimum character to fetch results
